restructured so the web api backend can run on it's own
