+++ FlotillaWeb/.dockerignore
@@ -1,0 +1,5 @@
+Dockerfile
+*.md
+*/bin
+dockerfile
+FlotillaWeb
