+++ FlotillaWebAPI/dockerfile
@@ -1,0 +1,10 @@
+FROM ximidar/flotilla_root:latest
+
+WORKDIR $FLOTILLA_DIR/FlotillaWebAPI/
+COPY . .
+RUN go build -ldflags "-X 'github.com/Ximidar/Flotilla/CommonTools/versioning.Version=0.0.1' -X 'github.com/Ximidar/Flotilla/CommonTools/versioning.CompiledBy=Matt Pedler' -X 'github.com/Ximidar/Flotilla/CommonTools/versioning.CompiledDate=`date '+%d %b %y at %H:%M:%S %p'`'" -o FlotillaWebAPI
+#RUN make
+
+RUN mkdir dist
+EXPOSE 5000
+CMD ./FlotillaWebAPI serve
