+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/Ximidar/Flotilla
+
+go 1.13
